Basic import af simple document editor
This is an impert of the simple document editor.  We have imported
ember-vo-webuniversum, postscribe and ember-concurrency, which
probably should be imported by dependencies in the future.  The style
is not yet what you'd expect of it, but that is likely something that
will need to be adjusted to the specific domain.