Add interface on the DOM node of the editor
This commit adds an interface to the DOM node of the editor.  We've
altered the tagName to notule-editor in order to make the editor
easier to discover.

The method to set the HTML content should be implemented correctly.
This will currently only work when the contents are set before the
editor is initialized which is not the ideal case.