Configure app for manual boot embedded in HTML page
Since the embeddable-notule-editor will be inserted in another HTML page
we need to prevent the router's location adapter to update the URL in
the address bar.
We also disable autoboot of the application on DOM ready such that the
root element to which the Ember application will be attached can be
configured per application instance before booting.
To facilitate the integration in an HTML page we also package the
configuration in the javascript package instead of a metatag and disable
the fingerprint on the generated assets.