Update of the README file
This version adds the new .js file to be imported, but also corrects
some spelling mistakes and overall lay-out improvements.
Another thing changed is that there is now one editor being initialised
instead of two. People generally don't need two editors on the same
page, and if so, the mechanism is easy enough so that developers can
duplicate it multiple times on the same page.