Fix release process to correctly bump embeddable version in test app
Note: the 'standard' release-it solution of using version-bumper to
bump the version produces a new copy of the pnpm-lock file with missing
newlines, which then results in pnpm recreating it. In order to avoid
constant flipping between these versions, we just use pnpm to increment
it. This potentially leaves the possibility of misconfigured lockfile
contents being fixed in the release process. The chances of this getting
through the build process before this are small though.